In crypto_hash() lengths are ULL, so dont loop using and int value

+++ src/libsodium/crypto_auth/hmacsha256/ref/hmac_hmacsha256.c
@@ -26,7 +26,7 @@ int crypto_auth(unsigned char *out,const unsigned char *in,unsigned long long in
 {
   unsigned char h[32];
   unsigned char padded[128];
-  int i;
+  unsigned long long i;
   unsigned long long bits = 512 + (inlen << 3);
 
   for (i = 0;i < 32;++i) h[i] = iv[i];
